添加一些任务

+++ HARDWARE/inc/motor.h
@@ -4,6 +4,6 @@
 #include "stm32f10x.h"
 
 void SetMotorLeftPower(int32_t power);
-void SetMotorRightPower(uint16_t power);
+void SetMotorRightPower(int32_t power);
 
 #endif // __MOTOR_H_
